requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
passlib==1.7.4
PyJWT==2.8.0
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.11.0